requirements.txt
asgiref==3.4.1
cffi==1.15.1
cryptography==37.0.2
Django==3.2.13
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
django-rest-knox==4.1.0
djangorestframework==3.13.1
phonenumbers==8.12.50
psycopg2-binary==2.9.3
pycparser==2.21
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
